+++ aws-cdk/cdk-hello-world/stack.ts
